Appointment services changes

+++ Covid_19_Vaccination/src/main/java/com/masaischool/service/SerImpl/AppointmentServiceImpl.java
@@ -1,11 +1,15 @@
 package com.masaischool.service.SerImpl;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.masaischool.entity.*;
 import com.masaischool.exception.AppointmentNotFoundException;
 import com.masaischool.repository.AppointmentRepository;
+import com.masaischool.repository.MemberRepository;
+import com.masaischool.repository.VaccinationCenterRepository;
 import com.masaischool.service.AppointmentService;
 
 @Service
@@ -16,6 +20,11 @@ public class AppointmentServiceImpl implements AppointmentService{
 	@Autowired 
 	private AppointmentRepository appointmentRepository;
 	
+	@Autowired
+	private MemberRepository memberRepository;
+	
+	@Autowired
+	private VaccinationCenterRepository vaccinationCenterRepository;
 
 	@Override
 	public Appointment getAppointmentDetails(Integer bookingId) {
@@ -25,36 +34,38 @@ public class AppointmentServiceImpl implements AppointmentService{
 
 	@Override
 	public Appointment addAppointment(Integer memberId, Integer vaxCenterId, Appointment appointment) {
-		 Member member = new Member();
-	     member.setId(memberId);
-	     VaccinationCenter vaccinationCenter = new VaccinationCenter();
-	     vaccinationCenter.setCenterId(vaxCenterId);
-
-	     appointment.setMemberId(member);
-	     appointment.setVaxCenter(vaccinationCenter);
-
-	     return appointmentRepository.save(appointment);
+    if(appointment == null) throw new AppointmentNotFoundException("Appointment couldn't saved");
+		
+		Member member = memberRepository.findById(memberId).orElseThrow(()-> new AppointmentNotFoundException("Please Provide a valid MemberId to book an Appointment"));
+		VaccinationCenter vaxCenter = vaccinationCenterRepository.findById(vaxCenterId).orElseThrow(()-> new AppointmentNotFoundException("Please Provide a valid Vaccination center id to book an Appointment"));
+		
+		member.setAppointment(appointment);
+		
+		vaxCenter.getAppointments().add(appointment);
+		
+		appointment.setMemberId(member);
+		appointment.setVaxCenter(vaxCenter);
+		appointment.setSlot(Slot.SLOT1);
+		Appointment appoint = appointmentRepository.save(appointment);
+		return appoint;
 	}
 
 	@Override
 	public Appointment updateAppointment(Integer bookingId, Appointment appointment) {
-		Appointment existingAppointment = appointmentRepository.findById(bookingId)
-                .orElseThrow(() -> new AppointmentNotFoundException("Appointment not found with ID: " + bookingId));
-
-        existingAppointment.setDateOfBooking(appointment.getDateOfBooking());
-        existingAppointment.setSlot(appointment.getSlot());
-        // Update other fields as needed
-
-        return appointmentRepository.save(existingAppointment);
+		if(bookingId==null) throw new AppointmentNotFoundException("Please provide Appointment id you want to update");
+		if(appointment==null) throw new AppointmentNotFoundException("Please provide Updated Appointment details");
+		appointment.setBookingId(bookingId);
+		return appointmentRepository.save(appointment);
 	}
 
 	@Override
 	public Appointment deleteAppointment(Integer bookingId) {
-		Appointment appointment = appointmentRepository.findById(bookingId)
-                .orElseThrow(() -> new AppointmentNotFoundException("Appointment not found with ID: " + bookingId));
-
-        appointmentRepository.delete(appointment);
-        return appointment;
+		if(bookingId==null) throw new AppointmentNotFoundException("Please provide Appointment id you want to delete");
+		Optional<Appointment> optional = appointmentRepository.findById(bookingId);
+		if(optional.isEmpty()) throw new AppointmentNotFoundException("Appointment Not Found");
+		Appointment appointment = optional.get();
+		appointmentRepository.delete(appointment);
+		return appointment;
 	}
 
 }
